+++ apps/website/astro.config.mjs
@@ -1,6 +1,8 @@
+import sitemap from '@astrojs/sitemap'
 import { defineConfig } from 'astro/config'
 
 export default defineConfig({
+  integrations: [sitemap()],
   output: 'static',
   site: 'https://difftale.santi020k.com'
 })